alx_travel_app/requirement.txt
celery==5.5.2
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10		# For Swagger documentation
mysqlclient==2.2.7
